+++ src/main/java/net/sf/jabref/sql/DatabaseUtil.java
@@ -34,30 +34,25 @@ public class DatabaseUtil {
      * ID for this new database
      *
      * @param databaseContext the database
-     * @param out The output (PrintStream or Connection) object to which the DML should be written.
+     * @param out             The output (PrintStream or Connection) object to which the DML should be written.
      * @return The ID of database row of the jabref database being exported
      * @throws SQLException
      */
-    public static int getDatabaseIDByName(BibDatabaseContext databaseContext, Object out, String dbName) throws SQLException {
-        if (out instanceof Connection) {
-            String query = "SELECT database_id FROM jabref_database WHERE database_name='" + dbName + "';";
-            try (Statement statement = (Statement) ((Connection) out).createStatement();
-                 ResultSet rs = statement.executeQuery(query)) {
-                if (rs.next()) {
-                    return rs.getInt("database_id");
-                } else {
-                    insertJabRefDatabase(databaseContext, out, dbName);
-                    return getDatabaseIDByName(databaseContext, out, dbName);
-                }
+    public static int getDatabaseIDByName(BibDatabaseContext databaseContext, Connection out, String dbName)
+            throws SQLException {
+        String query = "SELECT database_id FROM jabref_database WHERE database_name='" + dbName + "';";
+        try (Statement statement = (Statement) ((Connection) out).createStatement();
+                ResultSet rs = statement.executeQuery(query)) {
+            if (rs.next()) {
+                return rs.getInt("database_id");
+            } else {
+                insertJabRefDatabase(databaseContext, out, dbName);
+                return getDatabaseIDByName(databaseContext, out, dbName);
             }
-        } else {
-            // in case of text export there will be only 1 bib exported
-            insertJabRefDatabase(databaseContext, out, dbName);
-            return 1;
         }
     }
 
-    private static void removeAGivenDB(Object out, final int database_id) throws SQLException {
+    private static void removeAGivenDB(Connection out, final int database_id) throws SQLException {
         removeAllRecordsForAGivenDB(out, database_id);
         SQLUtil.processQuery(out, "DELETE FROM jabref_database WHERE database_id='" + database_id + "';");
     }
@@ -65,11 +60,11 @@ public class DatabaseUtil {
     /**
      * Removes all records for the database being exported in case it was exported before.
      *
-     * @param out The output (PrintStream or Connection) object to which the DML should be written.
+     * @param out         The output (PrintStream or Connection) object to which the DML should be written.
      * @param database_id Id of the database being exported.
      * @throws SQLException
      */
-    public static void removeAllRecordsForAGivenDB(Object out, final int database_id) throws SQLException {
+    public static void removeAllRecordsForAGivenDB(Connection out, final int database_id) throws SQLException {
         SQLUtil.processQuery(out, "DELETE FROM entries WHERE database_id='" + database_id + "';");
         SQLUtil.processQuery(out, "DELETE FROM groups WHERE database_id='" + database_id + "';");
         SQLUtil.processQuery(out, "DELETE FROM strings WHERE database_id='" + database_id + "';");
@@ -79,11 +74,11 @@ public class DatabaseUtil {
      * This method creates a new row into jabref_database table enabling to export more than one .bib
      *
      * @param databaseContext the database
-     * @param out The output (PrintStream or Connection) object to which the DML should be written.
-     *
+     * @param out             The output (PrintStream or Connection) object to which the DML should be written.
      * @throws SQLException
      */
-    private static void insertJabRefDatabase(final BibDatabaseContext databaseContext, Object out, String dbName) throws SQLException {
+    private static void insertJabRefDatabase(final BibDatabaseContext databaseContext, Connection out, String dbName)
+            throws SQLException {
         String path;
         if (databaseContext.getDatabaseFile() == null) {
             path = dbName;