+++ CHANGELOG.md
@@ -22,9 +22,12 @@ to [sourceforge feature requests](https://sourceforge.net/p/jabref/features/) by
 - Added integrity check detecting HTML-encoded characters
 
 ### Fixed
+- Fixed [#473](https://github.com/JabRef/jabref/issues/473): Values in an entry containing symbols like ' are now properly escaped for exporting to the database
+
 - Fixed [#1234](https://github.com/JabRef/jabref/issues/1234): NPE when getting information from retrieved DOI
 
 ### Removed
+- Removes possibility to export entries/databases to an .sql file, as the logic cannot easily use the correct escape logic